requirements.txt
yfinance
pandas
duckduckgo-search
google-generativeai
langchain
langchain-community
schedule
streamlit
plotly
